Fix Postgres-isms that only surfaced when the app actually ran
Running the app against local D1 found three classes of leftover Postgres
syntax that grepping had missed, because my earlier sweep only matched
single-line sql`` templates and these all lived in multi-line ones.

  - 18 `::int` casts, mostly `COUNT(*)::int` inside correlated subqueries.
    SQLite has no cast-colon syntax at all, so these were hard parse
    errors: `unrecognized token: ":"`. /rip and the whole admin surface
    500'd on them. count() already returns an integer, so the casts just
    go away.

  - Two more `sql\`now()\`` sites in ingest.ts and user-packs.ts, now
    `new Date()` like the rest.

  - Two JS Dates interpolated straight into raw sql`` templates
    (`packRips.rippedAt > ${weekAgo}`). This is the failure mode already
    documented for social.ts's cursor params, but I had only fixed the
    three sites I could see. D1 accepts null/number/string/boolean/
    ArrayBuffer and rejects a Date outright.

The Date helper moves to src/db/sql.ts so there is one copy rather than a
per-file one, and a script now checks the whole tree for Dates bound into
raw templates — that check is what found these two.

Verified against local D1 by running the queries, not just by reading
them. The rewritten social aggregations were the ones worth proving, so
they were exercised against a synthetic rip: json_each correctly unnests
the JSON id array, json_group_array builds one row per rip, ORDER BY
inside the derived table sorts the cards, and `json(b.authors)` embeds
authors as a real array rather than an escaped string — the specific trap
in that rewrite. json_array_length replaces array_length correctly.

The compare-and-swap debit was exercised both ways: it succeeded at 50
against a balance of 52, then refused a second 50 against the remaining
2, and the paired conditional INSERT ... SELECT wrote zero rows under the
same predicate. That is the property the old SELECT ... FOR UPDATE was
there to provide. Synthetic rows were removed afterwards; local D1 is back
to balance 52, ledger 52, 0 rips, 23 events.

All routes now respond (200, or 307 to sign-in for authenticated ones)
with zero query failures in the Worker log.

+++ src/lib/packs/unlock.ts
@@ -40,10 +40,10 @@ export async function getPublishUnlockStatus(
 ): Promise<PublishUnlockStatus> {
   const [cfg, database] = await Promise.all([getEconomy(), getDb()]);
   // Count, not row fetch — we only need the number and the index on
-  // (user_id, status, updated_at) makes this cheap. Uses `count(*)::int`
+  // (user_id, status, updated_at) makes this cheap. Uses `count(*)`
   // so drizzle hands us a plain number rather than a string.
   const [row] = await database
-    .select({ n: sql<number>`count(*)::int` })
+    .select({ n: sql<number>`count(*)` })
     .from(readingEntries)
     .where(
       and(
